Added preprocessing for data from KBR exports to fix #3
Added the script  to adapt the heades of an input CSV File. The script uses two csv files, the first one is the input and the second one is a 'mapping' between old header and new header; a third csv file is created which is then the output. See  for more information