use different source fields for publishers and try to avoid duplicates. A hash value is generated as ID for publishers without authority record to fix #34
We consider the MARC field 710, but we also check the field 264 for publishers. Publishers are always in 264 but not always in 710. Thus there might be doubles. We check with string comparison, if one is a substring of the other, or based on the levenshtein difference of both names (word by word if necessary)