create stats including BnF to fix #72
The stats are now generated with Python Pandas based on the CSV files we query via SPARQL. BnF is therefore included and the statistics are now 'combined' and 'single', i.e. If a book was published in 'Dutch,English,Spanish' it is counted for each of the language once in the 'single' output but only once in the combined stats for the unique value 'Dutch,English,Spanish'.