draft of filtering possible matches #129
Promising results, but on the one hand the code needs to be refactored (currently not following the DRY principle) and on the other hand we need to figure out how to deal with cases where we basically filter out all possible solutions, i.e. all found originals are reprints with a later publishing year than the translation. We would end up with no link instead of at least a link to a wrong edition, but correct original