added refactored code to lookup KB publisher data in XML files #152 #117
copied the original get-subjects.py and filter-subjects-xml.py scripts from data-sources/bnf and adapted it with a new parameter to choose the resource of the subject, e.g. before it was always rdf:Description, but for KB data it is schema:Organization. This is now configurable via parameter. The script needs to be called with python -m tools.xml.get-subjects