Added more information about publishers to fix #209
The sheet org-contributors now also contains statistics about how often the org published a translation or an original (the original count is 0 for now). Additionally, all organizations are listed in the new sheet all-orgs. The post-process script was adapted to also handle organizations. The changes are a bit hacky, the script would profit from some refactoring to be more modular, or by splitting the script into two scripts reusing modular functions.